Move tests to where they belong and ensure all namespaces match the path.
