+++ tripal_chado/tests/src/Functional/Services/ChadoTripalPublishTest.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Drupal\Tests\tripal_chado\Functional\Service;
+namespace Drupal\Tests\tripal_chado\Functional\Services;
 
 use Drupal\Tests\tripal_chado\Functional\ChadoTestBrowserBase;
 use Drupal\tripal\TripalVocabTerms\TripalTerm;
