+++ tripal_chado/tests/src/Kernel/Plugin/ChadoField/FieldType/ChadoPropertyTypeCRUDTest.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Drupal\Tests\tripal_chado\Kernel\ChadoField\FieldType;
+namespace Drupal\Tests\tripal_chado\Kernel\Plugin\ChadoField\FieldType;
 
 use Drupal\Tests\tripal_chado\Kernel\ChadoTestKernelBase;
 use Drupal\Tests\tripal_chado\Traits\ChadoFieldTestTrait;
