+++ tripal_chado/tests/src/Kernel/Plugin/ChadoField/Widget/ChadoPropertyWidgetFormTest.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Drupal\Tests\tripal_chado\Kernel\ChadoField\FieldType;
+namespace Drupal\Tests\tripal_chado\Kernel\Plugin\ChadoField\Widget;
 
 use Drupal\Tests\tripal_chado\Kernel\ChadoTestKernelBase;
 use Drupal\Tests\tripal_chado\Traits\ChadoFieldTestTrait;
